+++ NEWS.md
@@ -1,5 +1,12 @@
 # 0.4.9 ongoing
+* Removed suggestion of `relax` package (see below).
+
+* `assumptionCheck()`: Modified. Removed option to use `relax()` and Tcl/Tk version of the dynamic graphic.
+* `ciSim()`: Modified. Removed option to use `relax()` and Tcl/Tk version of the dynamic graphic.
+* `cltSim()`: Modified. Removed option to use `relax()` and Tcl/Tk version of the dynamic graphic.
 * `filterD()`: Added (from `FSA`).
+* `meanMedian()`: Modified. Removed option to use `relax()` and Tcl/Tk version of the dynamic graphic.
+* `powerSim()`: Modified. Removed option to use `relax()` and Tcl/Tk version of the dynamic graphic.
 
 # 0.4.8 15-Jan-21
 * Changed to depending on `R >=3.5.0`, because that is the latest version required by packages (i.e., `car`, `FSA`, `plotrix`) that `NCStats` imports or suggests. Used [this shiny app](https://ateucher.shinyapps.io/check_r_versions_of_package_dependencies/) to help determine this.